requirements.txt
colorama==0.4.6
Pillow==9.3.0
pyotp==2.7.0
PyQt6==6.4.0
PyQt6-Qt6==6.4.1
PyQt6-sip==13.4.0
qrcode==7.3.1
